refactor: locate update user logic to services
