+++ controllers/adminController.js
@@ -68,74 +68,18 @@ export const createUser = async (req, res) => {
     }
 }
 
-export const putUser = async (req, res) => {
+export const updateUser = async (req, res) => {
     try {
-        const userId = parseInt(req.params.user_id);
+        const userId = req.params.user_id;
 
-        if (isNaN(userId)) {
-          console.log('Invalid user ID format');
-          return res.status(400).json({ error: 'Invalid user ID format' });
-        }
+        const result = await adminService.updateUserData(userId, req.body);
+        return res.status(200).json(result);
 
-        const userData = await User.getUserData(userId);
-
-        if (!userData) {
-          console.log('No user found for ID:', userId);
-          return res.status(404).json({ error: 'No information found for the user' });
-        }
-
-        const newUserData = req.body;
-        const isEmail = await Admin.findUserByEmail(req.body.email);
-        if (isEmail) {
-            return res.status(400).json({error: "Email already in use"});
-        }
-        const updatedFields = [];
-        const fieldsToUpdateInDb = {};
-        const keysToCompare = [
-            'role_name',
-            'department_name',
-            'user_name',
-            'workstation',
-            //'password',
-            'email',
-            'phone_number',
-        ];
-
-        for (const key of keysToCompare) {
-            if (newUserData[key] !== undefined && newUserData[key] !== userData[key]) {
-                updatedFields.push(key);
-                if (key === 'role_name') {
-                    const roleID = await Admin.findRoleID(newUserData[key]);
-                    if (roleID !== null) {
-                        fieldsToUpdateInDb.role_id = roleID;
-                    } else {
-                         return res.status(400).json({ error: `Invalid role name provided: ${newUserData[key]}` });
-                    }
-                } else if (key === 'department_name') {
-                    const deptId = await Admin.findDepartmentID(newUserData[key]);
-                     if (deptId !== null) {
-                         fieldsToUpdateInDb.department_id = deptId;
-                     } else {
-                          return res.status(400).json({ error: `Invalid department name provided: ${newUserData[key]}` });
-                     }
-                } else {
-                    fieldsToUpdateInDb[key] = newUserData[key];
-                }
-            }
-        }
-
-        if (Object.keys(fieldsToUpdateInDb).length > 0) {
-            await Admin.updateUser(userId, fieldsToUpdateInDb);
-
-            return res.status(200).json({ message: "User updated successfully", updated_fields: updatedFields });
-        } else {
-            return res.status(200).json({ message: "No changes detected, user data is up to date" });
-        }
     } catch (error) {
-        console.error('An error occurred in putUser:', error);
-        return res.status(500).json({description: "Internal server error"});
+        console.error('An error occurred updating the user:', error);
+        return res.status(error.status || 500).json({ error: error.message || 'Internal server error' });
     }
-}
+};
 
 export const deactivateUser = async (req, res) => {
     try {
