+++ routes/adminRoutes.js
@@ -9,8 +9,6 @@ import { authenticateToken, authorizeRole } from "../middleware/auth.js";
 import { validateCreateUser, validateInputs } from "../middleware/validation.js";
 import { generalRateLimiter } from "../middleware/rateLimiters.js";
 
-import { getUserList, putUser } from "../controllers/adminController.js";
-
 const upload = multer({
     dest: "uploads/"
 });
@@ -34,7 +32,7 @@ router.route("/create-multiple-users")
     );
 
 router.route('/update-user/:user_id')
-    .put(putUser);
+    .put(adminController.updateUser);
 
 router.route("/delete-user/:user_id")
     .put(adminController.deactivateUser);
